fix(tag): [tag] Fix the pre deletion operation function
