+++ examples/sites/demos/pc/app/tag/delete-composition-api.vue
@@ -1,16 +1,11 @@
 <template>
   <div>
-    <p>
-      <tiny-switch v-model="beforeDeleteValue"></tiny-switch>
-      before-change 返回 {{ beforeDeleteValue }} 时 {{ beforeDeleteText }}
-      <br />
-    </p>
     <tiny-tag
       v-for="tag in tags"
       :key="'tiny-tag1-' + tag.name"
       closable
       :type="tag.type"
-      @close="close(tag)"
+      @close="handleClose(tag)"
       :before-delete="beforeDelete"
       >{{ tag.name }}</tiny-tag
     >
@@ -19,7 +14,7 @@
 
 <script setup>
 import { ref, computed } from 'vue'
-import { Tag as TinyTag, Switch as TinySwitch } from '@opentiny/vue'
+import { Tag as TinyTag, Modal } from '@opentiny/vue'
 
 const tags = ref([
   { name: '标签一', type: '' },
@@ -28,20 +23,15 @@ const tags = ref([
   { name: '标签四', type: 'warning' },
   { name: '标签五', type: 'danger' }
 ])
-const beforeDeleteValue = ref(false)
-const beforeDeleteText = computed(() => {
-  return beforeDeleteValue.value
-    ? '可执行 close 事件，点击删除图标可以删除标签'
-    : '将拦截 close 事件的执行，点击删除图标将无法删除标签'
-})
 
-function close(tag) {
+function handleClose(tag) {
   tags.value.splice(tags.value.indexOf(tag), 1)
+  Modal.message('close 事件')
 }
 
 function beforeDelete(done) {
-  setTimeout(() => {
-    beforeDeleteValue.value && done()
+  Modal.confirm('确认删除？').then((res) => {
+    res === 'confirm' && done()
   })
 }
 </script>
